Split viva.js into a bunch of small modules
